On create project task added to sqs and db
