+++ client/lib/aws.js
@@ -1,0 +1,9 @@
+import { SQSClient } from "@aws-sdk/client-sqs";
+
+export const sqsClient = new SQSClient({
+  region: process.env.AWS_REGION,
+  credentials: {
+    accessKeyId: process.env.AWS_ACCESS_KEY,
+    secretAccessKey: process.env.AWS_SECRET_KEY,
+  },
+});
