+++ client/lib/octokit.js
@@ -1,0 +1,5 @@
+import { Octokit } from "@octokit/rest";
+
+export const octokit = new Octokit({
+  auth: process.env.GITHUB_ACCESS_TOKEN,
+});
